backend/requirements.txt
fastapi
uvicorn[standard]
bcrypt
python-jose
pydantic
python-multipart